requirements.txt
python-dotenv
rdflib
requests

